Rework on the veracruz-server-test.
